requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0

# Original dependencies
python-dotenv==1.0.0
requests==2.31.0
llama-api-client

# Optional: For production deployment
gunicorn==21.2.0
